pdf-service/requirements.txt
aws-cdk-lib==2.180.0
constructs>=10.0.0,<11.0.0
boto3==1.37.4
botocore==1.37.4
fire==0.7.0
fonttools==4.56.0
jmespath==1.0.1
lxml==5.3.1
numpy==2.2.3
opencv-python-headless==4.11.0.86
pdf2docx==0.5.8
PyMuPDF==1.25.3
python-dateutil==2.9.0.post0
python-docx==1.1.2
s3transfer==0.11.3
six==1.17.0
termcolor==2.5.0
typing_extensions==4.12.2
urllib3==2.3.0
